refactor: Move `ConversationInfoView` out of `ConversationFeature` to benefit from namespacing
