+++ Prose/ProseLib/Sources/ConversationFeature/ConversationScreen.swift
@@ -6,7 +6,9 @@
 //
 
 import ComposableArchitecture
+import ConversationInfoFeature
 import PreviewAssets
+import ProseCoreStub
 import SharedModels
 import SwiftUI
