+++ Prose/ProseLib/Sources/ConversationFeature/Toolbar/Toolbar.swift
@@ -6,6 +6,7 @@
 //
 
 import ComposableArchitecture
+import ProseCoreStub
 import ProseUI
 import SwiftUI
